requirements.txt
gTTS==2.5.3
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
ollama==0.3.3
openai-whisper
python-dotenv==1.0.1
PyYAML==6.0.2
sounddevice==0.5.0
soundfile==0.12.1
streamlit==1.38.0
streamlit-webrtc==0.47.9
streamlit-authenticator
torch==2.3.1